requirements.txt
pandas>=2.0
openpyxl>=3.1
rapidfuzz>=3.0
pdfplumber>=0.10
typer[all]>=0.9
python-dotenv>=1.0
requests>=2.31
pytest>=7.0
sqlalchemy>=2.0
